SafeDrive/reward_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
